# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Project Heartstar API
  version: 0.3.0
  description: |
    Pixel adventure API contract.
    Auth uses username + 4-digit PIN and bearer token sessions.
servers:
  - url: https://api.example.com
...
paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with username and 4-digit PIN
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and invalidate current session
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    AuthInput:
      type: object
      required: [username, pin]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 24
          pattern: "^[A-Za-z0-9_]+$"
        pin:
          type: string
          description: Exactly 4 digits
          pattern: "^\\d{4}$"
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
    PlayerProgress:
      type: object
      required: [totalHearts, totalStarsLit, bestCombo, updatedAt]
      properties:
        totalHearts:
          type: integer
          minimum: 0
        totalStarsLit:
          type: integer
          minimum: 0
        bestCombo:
          type: integer
          minimum: 0
        updatedAt:
          type: string
    GameState:
      type: object
      required: [coin, progress]
      properties:
        coin:
          type: integer
          minimum: 0
        progress:
          $ref: "#/components/schemas/PlayerProgress"
    AuthSuccess:
      type: object
      required: [accessToken, user, state]
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
        state:
          $ref: "#/components/schemas/GameState"
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_INPUT
        message:
          type: string
          example: Invalid username format.